+++ Fluent.Ribbon/Adorners/KeyTipAdorner.cs
@@ -1,19 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Windows;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Interop;
-using System.Windows.Media;
-using System.Windows.Threading;
-
-namespace Fluent
+﻿namespace Fluent
 {
+    using System;
+    using System.Collections.Generic;
     using System.Diagnostics;
+    using System.Diagnostics.CodeAnalysis;
+    using System.Linq;
+    using System.Windows;
     using System.Windows.Controls;
+    using System.Windows.Data;
+    using System.Windows.Documents;
+    using System.Windows.Media;
     using Fluent.Internal;
 
     /// <summary>
@@ -46,23 +42,14 @@ namespace Fluent
         private readonly KeyTipAdorner parentAdorner;
         KeyTipAdorner childAdorner;
 
-        // Focused element
-        private IInputElement focusedElement;
-
         private readonly Visibility[] backupedVisibilities;
         private readonly UIElement keyTipElementContainer;
 
         // Is this adorner attached to the adorned element?
         private bool attached;
-        private HwndSource attachedHwndSource;
-
-        // Current entered chars
-        string enteredKeys = "";
 
         // Designate that this adorner is terminated
         private bool terminated;
-
-        private DispatcherTimer timerFocusTracking;
 
         private AdornerLayer adornerLayer;
 
@@ -245,92 +232,14 @@ namespace Fluent
                 return;
             }
 
-            // Focus current adorned element
-            // Keyboard.Focus(adornedElement);
-            this.focusedElement = Keyboard.FocusedElement;
-
-            if (this.focusedElement != null)
-            {
-                this.Log("Focus Attached to {0}", this.focusedElement);
-                this.focusedElement.LostKeyboardFocus += this.OnFocusLost;
-                this.focusedElement.PreviewKeyDown += this.OnPreviewKeyDown;
-                this.focusedElement.PreviewTextInput += this.OnFocusedElementPreviewTextInput;
-            }
-            else
-            {
-                this.Log("[!] Focus Setup Failed");
-            }
-
-            GetTopLevelElement(this.oneOfAssociatedElements).PreviewMouseDown += this.OnInputActionOccured;
-
-            // Clears previous user input
-            this.enteredKeys = "";
-            this.FilterKeyTips(this.enteredKeys);
+            this.FilterKeyTips(string.Empty);
 
             // Show this adorner
             this.adornerLayer.Add(this);
 
-            // Hookup window activation
-            this.attachedHwndSource = ((HwndSource)PresentationSource.FromVisual(this.oneOfAssociatedElements));
-            if (this.attachedHwndSource != null)
-            {
-                this.attachedHwndSource.AddHook(this.WindowProc);
-            }
-
-            // Start timer to track focus changing
-            if (this.timerFocusTracking == null)
-            {
-                this.timerFocusTracking = new DispatcherTimer(DispatcherPriority.ApplicationIdle, Dispatcher.CurrentDispatcher)
-                {
-                    Interval = TimeSpan.FromMilliseconds(50)
-                };
-                this.timerFocusTracking.Tick += this.OnTimerFocusTrackingTick;
-            }
-
-            this.timerFocusTracking.Start();
-
             this.attached = true;
 
             this.Log("Attach end");
-        }
-
-        private void OnTimerFocusTrackingTick(object sender, EventArgs e)
-        {
-            if (this.focusedElement == Keyboard.FocusedElement)
-            {
-                return;
-            }
-
-            this.Log("Focus is changed, but focus lost is not occured");
-
-            if (this.focusedElement != null)
-            {
-                this.focusedElement.LostKeyboardFocus -= this.OnFocusLost;
-                this.focusedElement.PreviewKeyDown -= this.OnPreviewKeyDown;
-                this.focusedElement.PreviewTextInput -= this.OnFocusedElementPreviewTextInput;
-            }
-
-            this.focusedElement = Keyboard.FocusedElement;
-
-            if (this.focusedElement != null)
-            {
-                this.focusedElement.LostKeyboardFocus += this.OnFocusLost;
-                this.focusedElement.PreviewKeyDown += this.OnPreviewKeyDown;
-                this.focusedElement.PreviewTextInput += this.OnFocusedElementPreviewTextInput;
-            }
-        }
-
-        // Window's messages hook up
-        private IntPtr WindowProc(IntPtr hwnd, int msg, IntPtr wParam, IntPtr lParam, ref bool handled)
-        {
-            // Check whether window is deactivated (wParam == 0)
-            if ((msg == 6) && (wParam == IntPtr.Zero) && (this.attached))
-            {
-                this.Log("The host window is deactivated, keytips will be terminated");
-                this.Terminate();
-            }
-
-            return IntPtr.Zero;
         }
 
         private void OnDelayAttach(object sender, EventArgs args)
@@ -357,34 +266,13 @@ namespace Fluent
 
             this.Log("Detach Begin");
 
-            // Remove window hookup
-            if ((this.attachedHwndSource != null) && (!this.attachedHwndSource.IsDisposed))
-            {
-                // Crashes in a few time if invoke immediately ???
-                this.AdornedElement.Dispatcher.BeginInvoke((System.Threading.ThreadStart)(() => this.attachedHwndSource.RemoveHook(this.WindowProc)));
-            }
-
             // Maybe adorner awaiting attaching, cancel it
             this.oneOfAssociatedElements.Loaded -= this.OnDelayAttach;
 
-            if (this.focusedElement != null)
-            {
-                this.focusedElement.LostKeyboardFocus -= this.OnFocusLost;
-                this.focusedElement.PreviewKeyDown -= this.OnPreviewKeyDown;
-                this.focusedElement.PreviewTextInput -= this.OnFocusedElementPreviewTextInput;
-                this.focusedElement = null;
-            }
-
-            GetTopLevelElement(this.oneOfAssociatedElements).PreviewMouseDown -= this.OnInputActionOccured;
-
             // Show this adorner
             this.adornerLayer.Remove(this);
-            // Clears previous user input
-            this.enteredKeys = "";
-            this.attached = false;
 
-            // Stop timer to track focus changing
-            this.timerFocusTracking.Stop();
+            this.attached = false;
 
             this.Log("Detach End");
         }
@@ -407,116 +295,13 @@ namespace Fluent
 
             this.Detach();
 
-            if (this.parentAdorner != null)
-            {
-                this.parentAdorner.Terminate();
-            }
+            this.parentAdorner?.Terminate();
 
-            if (this.childAdorner != null)
-            {
-                this.childAdorner.Terminate();
-            }
+            this.childAdorner?.Terminate();
 
-            if (this.Terminated != null)
-            {
-                this.Terminated(this, EventArgs.Empty);
-            }
+            this.Terminated?.Invoke(this, EventArgs.Empty);
 
             this.Log("Termination");
-        }
-
-        #endregion
-
-        #region Event Handlers
-
-        [SuppressMessage("Microsoft.Maintainability", "CA1502")]
-        private void OnPreviewKeyDown(object sender, KeyEventArgs e)
-        {
-            this.Log("Key Down {0} ({1})", e.Key, e.OriginalSource);
-
-            if (e.IsRepeat
-                || this.Visibility == Visibility.Hidden)
-            {
-                return;
-            }
-
-            if ((!(this.AdornedElement is ContextMenu) && (!(this.AdornedElement is MenuItem))) &&
-                ((e.Key == Key.Left) || (e.Key == Key.Right) || (e.Key == Key.Up) || (e.Key == Key.Down) ||
-                (e.Key == Key.Enter) || (e.Key == Key.Tab)))
-            {
-                this.Visibility = Visibility.Hidden;
-            }
-            else if (e.Key == Key.Escape)
-            {
-                this.Back();
-                e.Handled = true;
-            }
-        }
-
-        private void OnFocusedElementPreviewTextInput(object sender, TextCompositionEventArgs e)
-        {
-            var keyToSearch = this.enteredKeys + e.Text;
-
-            if (this.IsElementsStartWith(keyToSearch))
-            {
-                this.enteredKeys += e.Text;
-
-                var element = this.TryGetElement(this.enteredKeys);
-
-                if (element != null)
-                {
-                    this.Forward(element);
-                }
-                else
-                {
-                    this.FilterKeyTips(this.enteredKeys);
-                }
-
-                e.Handled = true;
-            }
-            else
-            {
-                System.Media.SystemSounds.Beep.Play();
-            }
-        }
-
-        private void OnInputActionOccured(object sender, RoutedEventArgs e)
-        {
-            if (!this.attached)
-            {
-                return;
-            }
-
-            this.Log("Input Action, Keystips will be terminated");
-            this.Terminate();
-        }
-
-        private void OnFocusLost(object sender, RoutedEventArgs e)
-        {
-            if (!this.attached)
-            {
-                return;
-            }
-
-            this.Log("Focus Lost");
-
-            var previousFocusedElementElement = this.focusedElement;
-            this.focusedElement.LostKeyboardFocus -= this.OnFocusLost;
-            this.focusedElement.PreviewKeyDown -= this.OnPreviewKeyDown;
-            this.focusedElement.PreviewTextInput -= this.OnFocusedElementPreviewTextInput;
-            this.focusedElement = Keyboard.FocusedElement;
-
-            if (this.focusedElement != null)
-            {
-                this.Log("Focus Changed from {0} to {1}", previousFocusedElementElement, this.focusedElement);
-                this.focusedElement.LostKeyboardFocus += this.OnFocusLost;
-                this.focusedElement.PreviewKeyDown += this.OnPreviewKeyDown;
-                this.focusedElement.PreviewTextInput += this.OnFocusedElementPreviewTextInput;
-            }
-            else
-            {
-                this.Log("Focus Not Restored");
-            }
         }
 
         #endregion
@@ -568,10 +353,7 @@ namespace Fluent
         public void Back()
         {
             var control = this.keyTipElementContainer as IKeyTipedControl;
-            if (control != null)
-            {
-                control.OnKeyTipBack();
-            }
+            control?.OnKeyTipBack();
 
             if (this.parentAdorner != null)
             {
@@ -606,15 +388,9 @@ namespace Fluent
         }
 
         // Forward to the next element
-        private void Forward(UIElement element)
-        {
-            this.Forward(element, true);
-        }
-
-        // Forward to the next element
         private void Forward(UIElement element, bool click)
         {
-            this.Log("Forwarding");
+            this.Log("Forwarding to {0}", element);
 
             this.Detach();
 
@@ -678,11 +454,10 @@ namespace Fluent
         }
 
         /// <summary>
-        /// Is one of the elements starts with the given chars
+        /// Determines if an of the keytips contained in this adorner start with <paramref name="keys"/>
         /// </summary>
-        /// <param name="keys"></param>
-        /// <returns></returns>
-        public bool IsElementsStartWith(string keys)
+        /// <returns><c>true</c> if any keytip start with <paramref name="keys"/>. Otherwise <c>false</c>.</returns>
+        public bool ContainsKeyTipStartingWith(string keys)
         {
             foreach (var keyTip in this.keyTips.Where(x => x.IsEnabled))
             {
@@ -698,9 +473,9 @@ namespace Fluent
         }
 
         // Hide / unhide keytips relative matching to entered keys
-        internal void FilterKeyTips(string currentlyEnteredKeys)
+        internal void FilterKeyTips(string keys)
         {
-            this.Log("FilterKeyTips");
+            this.Log("FilterKeyTips with {0}", keys);
 
             // Backup current visibility of key tips
             for (var i = 0; i < this.backupedVisibilities.Length; i++)
@@ -713,13 +488,13 @@ namespace Fluent
             {
                 var content = (string)this.keyTips[i].Content;
 
-                if (string.IsNullOrEmpty(currentlyEnteredKeys))
+                if (string.IsNullOrEmpty(keys))
                 {
                     this.keyTips[i].Visibility = this.backupedVisibilities[i];
                 }
                 else
                 {
-                    this.keyTips[i].Visibility = content.StartsWith(currentlyEnteredKeys, StringComparison.CurrentCultureIgnoreCase)
+                    this.keyTips[i].Visibility = content.StartsWith(keys, StringComparison.CurrentCultureIgnoreCase)
                         ? this.backupedVisibilities[i]
                         : Visibility.Collapsed;
                 }
@@ -959,25 +734,17 @@ namespace Fluent
                             elementSize.Height / 2.0 + 2,
                             elementSize.Height / 2.0 + 2), this.AdornedElement);
                 }
-                else if (this.associatedElements[i] is BackstageTabItem)
-                {
-                    // BackstageButton Exclusive Placement
-                    var keyTipSize = this.keyTips[i].DesiredSize;
-                    var elementSize = this.associatedElements[i].DesiredSize;
-                    this.keyTipPositions[i] = this.associatedElements[i].TranslatePoint(
-                        new Point(
-                            5,
-                            elementSize.Height / 2.0 - keyTipSize.Height), this.AdornedElement);
-                }
                 else if (((FrameworkElement)this.associatedElements[i]).Parent is BackstageTabControl)
                 {
                     // Backstage Items Exclusive Placement
                     var keyTipSize = this.keyTips[i].DesiredSize;
                     var elementSize = this.associatedElements[i].DesiredSize;
-                    this.keyTipPositions[i] = this.associatedElements[i].TranslatePoint(
+                    var parent = (UIElement)((FrameworkElement)this.associatedElements[i]).Parent;
+                    var positionInParent = this.associatedElements[i].TranslatePoint(default(Point), parent);
+                    this.keyTipPositions[i] = parent.TranslatePoint(
                         new Point(
-                            20,
-                            elementSize.Height / 2.0 - keyTipSize.Height), this.AdornedElement);
+                            5,
+                            positionInParent.Y + (elementSize.Height / 2.0 - keyTipSize.Height)), this.AdornedElement);
                 }
                 else
                 {
@@ -1110,10 +877,11 @@ namespace Fluent
             var headeredControl = this.AdornedElement as IHeaderedControl;
             if (headeredControl != null)
             {
-                name += string.Format(" ({0})", headeredControl.Header);
+                name += $" ({headeredControl.Header})";
             }
 
-            Debug.WriteLine("[" + name + "] " + string.Format(format, args), "KeyTipAdorner");
+            var formatted = string.Format(format, args);
+            Debug.WriteLine($"{$"[{name}] "}{formatted}", "KeyTipAdorner");
         }
 
         #endregion